python/requirements.txt
python==3.10.13
anndata==0.8.0
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
pyro_ppl==1.8.6
scanpy==1.9.3
scipy==1.9.3
seaborn==0.12.2
setuptools==70.0.0
torch==2.0.1
tqdm==4.66.3
upsetplot==0.8.0
dask>=2023.6.0
distributed>=2023.6.0